agentflow-app/backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
flaml
pyautogen<0.3
pydantic
crewai>=0.28.8
google-generativeai
litellm
requests
beautifulsoup4
lxml
python-telegram-bot>=20.0
